small fixes. implemented native clock
git-svn-id: https://midikit.googlecode.com/svn/trunk@72 c9dcb2c1-a6a4-760b-d948-6d6a4c2e33b0

+++ midi/midi.h
@@ -282,8 +282,8 @@
 typedef unsigned char  MIDIByte;
 typedef unsigned short MIDIProperty;
 
-typedef double    MIDISamplingRate;
-typedef long long MIDITimestamp;
+typedef unsigned int MIDISamplingRate;
+typedef long long    MIDITimestamp;
 
 typedef unsigned char  MIDIStatus;
 typedef unsigned char  MIDIRunningStatus;
